Move sglang config resolution into sglang_config.py
